Корректировка программы завершена

+++ src/main/java/org/example/Basket.java
@@ -28,7 +28,7 @@ public class Basket {
             Purchase purchase = purchases[i];
             if (purchase == null) continue;
             System.out.println("\t" + purchase.title + " " + purchase.count + " шт. в сумме " + (purchase.count * prices.get(purchase.title)) + " руб.");
-            sum += purchase.count * prices.get(purchase.title);
+            sum += (long) purchase.count * prices.get(purchase.title);
         }
         return sum;
     }
